add comment with logic

+++ src/main/kotlin/pl/put/smartgarden/domain/device/dto/response/MeasureResponse.kt
@@ -6,6 +6,6 @@ data class MeasureResponse(
     val id: Int,
     val timestamp: Instant,
     val sensorId: Int,
-    val value: Double,
+    val value: Int,
     val unit: String
 )
